requirements.txt
aiofiles==23.2.1
altair==5.3.0
amazon-textract-caller==0.0.27
amazon-textract-response-parser==0.1.27
amazon-textract-textractor==1.7.12
annotated-types==0.7.0
anthropic==0.26.1
anyio==4.4.0
attrs==23.2.0
bitarray==2.9.2
boto3==1.34.113
botocore==1.34.113
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
dnspython==2.6.1
editdistance==0.8.1
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.3
fonttools==4.53.0
fsspec==2024.6.0
gradio==4.37.2
gradio_client==1.0.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.11.1
matplotlib==3.9.0
mdurl==0.1.2
numpy==1.26.4
openpyxl==3.1.4
orjson==3.10.5
packaging==24.1
pandas==2.0.3
pdf2image==1.17.0
pdfreader==0.1.15
pillow==10.3.0
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.10
s3transfer==0.10.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tabulate==0.9.0
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==11.0.3
XlsxWriter==3.2.0
